+++ code/components/room-components/overlay/room-information-overlay.tsx
@@ -11,8 +11,10 @@ export function RoomInformationOverlay() {
     <div className="absolute top-2 left-2 flex gap-1 justify-center items-center">
       <div className="p-2 bg-white border border-zinc-200 shadow-xs flex justify-start items-center gap-2">
         <Logo kind="small" />
-        <div className="w-[1px] h-4 bg-zinc-200"></div>
-        <div className="flex justify-start items-center font-noto-sans-mono text-foreground !normal-case min-h-[32px]">{room}</div>
+        <div className="w-[1px] h-4 mx-2 bg-zinc-200"></div>
+        <div className="flex justify-start items-center font-noto-sans-mono text-foreground !normal-case min-h-[32px]">
+          {room}
+        </div>
       </div>
     </div>
   );
